Allow dividend pay/rel dates to be out of sequence
Dividend ex dates are sorted by definition. However, the pay dates are a settlement period after the ex date, and for some indices the settlement period can vary between the constituents of the index (EUROSTOXX 50 is an example).

We also allow the relative dates to be out of sequence, though this is less likely.